refactor: move logic for profile form into useFilledForm hook
